Place Login Buttons into LoginPanel Component
